refactor(batch-exports): move query range selection out of spmc
Second of three PRs emptying `spmc.py`.

Deciding which time ranges a run still needs to query, and which ClickHouse table
to read them from, moves to `pipeline/query_ranges.py`. Unlike filter composition
in the previous PR, this really is pipeline-specific: `internal_stage.py` is the
main caller, so it lands inside the package.

`slice_record_batch` and its helper fold into `pipeline/producer.py`, which was
their only production caller — a standalone module for two functions used from one
place wasn't worth it.

Both groups were self-contained: neither needed anything from what stays behind in
`spmc.py`, which is now down to 181 lines from 487. Code moved verbatim; all six
symbols verified byte-identical against master.

Generated-By: PostHog Code
Task-Id: e2c0e744-189b-4ac0-a1f4-c999b0a0e071
