+++ products/batch_exports/backend/temporal/pipeline/internal_stage.py
@@ -50,13 +50,13 @@ from products.batch_exports.backend.service import (
 from products.batch_exports.backend.temporal.batch_exports import default_fields
 from products.batch_exports.backend.temporal.filters import compose_filters_clause
 from products.batch_exports.backend.temporal.metrics import log_query_duration
-from products.batch_exports.backend.temporal.record_batch_model import RecordBatchModel, resolve_batch_exports_model
-from products.batch_exports.backend.temporal.spmc import (
+from products.batch_exports.backend.temporal.pipeline.query_ranges import (
     generate_query_ranges,
     is_5_min_batch_export,
     use_distributed_events_recent_table,
     wait_for_delta_past_data_interval_end,
 )
+from products.batch_exports.backend.temporal.record_batch_model import RecordBatchModel, resolve_batch_exports_model
 from products.batch_exports.backend.temporal.sql.common import get_s3_function_call
 from products.batch_exports.backend.temporal.sql.events import (
     EXPORT_TO_S3_FROM_DISTRIBUTED_EVENTS_RECENT,
